Storage Class: Coding quality of storage class
There could be higher level of abstraction for the parseTaskFromString as well as avoidance of magic literals

Created 3 new methods, createTodo, createDeadline, createEvent. As well as 4 fields to avoid magic literals.

Allow higher levels of abstraction as parseTaskFromString can directly call these methods when needed instead of implementing it. Allows more readability. The use of the constant fields also allows more readability to the code.